project restructure
add package.json
use global tsconfig
merge lockpick/laundromat code
fix some CSS
