+++ Laundromat/Laundromat.html
@@ -12,7 +12,7 @@
         <link rel="stylesheet" type="text/css" href="Laundromat.css">
         <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:400,400i,700&display=swap">
         <script src="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/js/all.min.js"></script>
-        <script src="Laundromat.js"></script>
+        <script type="module" src="Laundromat.js"></script>
     </head>
     <body>
         <div class="other-container">
