--- Laundromat/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "compilerOptions": {
-      "target": "es2015",
-      "module": "commonjs",
-      "strict": true,
-      "esModuleInterop": true,
-      "skipLibCheck": true,
-      "forceConsistentCasingInFileNames": true,
-      "outDir": "./"
-    },
-    "include": ["Laundromat.ts"],
-    "exclude": ["node_modules"]
-  }
-
