+++ LockPick/LockPick.css
@@ -34,7 +34,7 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    position: absolute; 
+    position: absolute;
     width: var(--box-size);
     height: var(--box-size);
 }
@@ -144,6 +144,7 @@
     background-color: orangered;
     width: 100%;
     height: 100%;
+    transition: width 1s linear;
 }
 
 /* The semi-circles that indicate where ball color is supposed to be */
@@ -167,11 +168,12 @@
 
 .lose-message {
     gap: 10px;
-    margin-top: -37px;
-    margin-left: 340px;
+    position: absolute;
+    right: 10px;
+    top: 10px;
     display: none;
     color: white;
-    padding: 0px 18px;
+    padding: 0px 17px;
     border-radius: 5px;
     align-items: center;
     justify-content: center;
@@ -180,8 +182,9 @@
 
 .win-message {
     gap: 10px;
-    margin-top: -40px;
-    margin-left: 340px;
+    position: absolute;
+    right: 10px;
+    top: 10px;
     display: none;
     color: white;
     padding: 0px 10px;
@@ -193,8 +196,9 @@
 
 .reset-message {
     gap: 10px;
-    margin-top: -40px;
-    margin-left: 500px;
+    position: absolute;
+    right: 10px;
+    top: 10px;
     display: none;
     color: white;
     padding: 0px 10px;
@@ -254,7 +258,13 @@
     -webkit-user-select: none;
     touch-action: manipulation;
   }
-  
+
+  .home-button:hover {
+    box-shadow: rgba(127, 255, 191, 0.397) 0 -25px 18px -14px inset,rgba(127, 255, 191, 0.397) 0 1px 2px,rgba(127, 255, 191, 0.397) 0 2px 4px,rgba(127, 255, 191, 0.397) 0 4px 8px,rgba(127, 255, 191, 0.397) 0 8px 16px,rgba(127, 255, 191, 0.397) 0 16px 32px;
+    transform: scale(1.05) rotate(-1deg);
+  }
+
+
   .home-button:hover {
     box-shadow: rgba(127, 255, 191, 0.397) 0 -25px 18px -14px inset,rgba(127, 255, 191, 0.397) 0 1px 2px,rgba(127, 255, 191, 0.397) 0 2px 4px,rgba(127, 255, 191, 0.397) 0 4px 8px,rgba(127, 255, 191, 0.397) 0 8px 16px,rgba(127, 255, 191, 0.397) 0 16px 32px;
     transform: scale(1.05) rotate(-1deg);
@@ -269,11 +279,11 @@
     right: 0px;
     transition: all 0.3s cubic-bezier(0.68, -0.55, 0.27, 1.55);
   }
-  
+
   #settings-icon:hover {
     transform: scale(1.1) rotate(270deg);
   }
-  
+
   .settings-container {
     flex-direction: column;
     background: radial-gradient(rgba(15, 27, 33, 0.781), rgb(15, 27, 33));
@@ -288,7 +298,7 @@
     gap: 50px;
     transform: translate(-50%, -50%);
   }
-  
+
   #settings-exit {
     position: absolute;
     color: rgb(94, 93, 93);
@@ -298,14 +308,14 @@
     height: 20px;
     cursor: pointer;
   }
-  
+
   .settings-header {
     font-size: 25px;
     color: rgb(84, 255, 164);
     text-shadow: 0 0 2.1px rgb(127, 255, 191);
     margin: 20px 0px 0px 20px;
   }
-  
+
   .sliders {
     color: rgb(94, 93, 93);
     font-size: 20px;
@@ -317,7 +327,7 @@
     width: 70%;
     justify-content: space-between;
   }
-  
+
   .timing-container {
     color: rgb(94, 93, 93);
     width: 50%;
@@ -329,12 +339,12 @@
     align-items: center;
     justify-content: center;
   }
-  
+
   .timing-container .slider-value {
     position: relative;
     width: 100%;
   }
-  
+
   .timing-container .slider-value span {
     position: absolute;
     text-align: center;
@@ -342,13 +352,13 @@
     width: 25px;
     color: rgba(22, 40, 52, 0.979);
     top: 60px;
-    left: 20%;
+    left: 12%;
     font-size: 15px;
     transform: translateX(-80%);
     line-height: 45px;
     z-index: 2;
   }
-  
+
   .timing-container .slider-value span:after {
     position: absolute;
     text-align: center;
@@ -365,7 +375,7 @@
     border-top-right-radius: 70%;
     z-index: -1;
   }
-  
+
   input {
     -webkit-appearance: none;
     appearance: none;
@@ -377,9 +387,9 @@
     border-radius: 10px;
     box-shadow: 0 0 2px rgb(127, 255, 191);
     transition: all 0.3s;
-  
+
   }
-  
+
   input::-moz-range-thumb {
     width: 20px;
     height: 20px;
@@ -389,7 +399,7 @@
     border: 1px solid rgba(22, 40, 52, 0.651);
     transition: all 0.3s;
   }
-  
+
   input::-moz-range-progress {
     width: 100%;
     height: 20px;
@@ -399,7 +409,7 @@
     border-bottom-left-radius: 10px;
     transition: all 0.3s;
   }
-  
+
   input::-webkit-slider-thumb {
     -webkit-appearance: none;
     appearance: none;
@@ -412,7 +422,7 @@
     transition: all 0.3s;
   }
 
-  
+
 .button-container {
   display: flex;
   align-items: center;
@@ -445,4 +455,53 @@
 #reset-button:hover {
   box-shadow: rgba(127, 255, 191, 0.397) 0 -25px 18px -14px inset,rgba(127, 255, 191, 0.397) 0 1px 2px,rgba(127, 255, 191, 0.397) 0 2px 4px,rgba(127, 255, 191, 0.397) 0 4px 8px,rgba(127, 255, 191, 0.397) 0 8px 16px,rgba(127, 255, 191, 0.397) 0 16px 32px;
   transform: scale(1.05) rotate(-1deg);
+}
+
+.countdown-container {
+  position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 10000;
+}
+
+.countdown {
+  position: absolute;
+  display: none;
+  font-size: 200px;
+  opacity: 0;
+  color: rgb(84, 255, 164);
+  text-shadow: 0 0 2.1px rgb(127, 255, 191);
+  margin: 0;
+  padding: 0;
+  animation: countdown-animation 900ms cubic-bezier(0.95, 0.05, 0.795, 0.035) forwards infinite;
+}
+
+.countdown.two {
+  animation-delay: 30ms;
+}
+.countdown.three {
+  animation-delay: 40ms;
+}
+.countdown.four {
+  animation-delay: 50ms;
+}
+.countdown.five {
+  animation-delay: 60ms;
+}
+
+@keyframes countdown-animation {
+  0% {
+    transform: scale(1);
+    opacity: 1;
+  }
+
+  80% {
+    transform: scale(1.5);
+    opacity: 0.4;
+  }
+  100% {
+    transform: scale(10);
+    opacity: 0;
+  }
 }
