+++ LockPick/LockPick.html
@@ -5,11 +5,14 @@
         <meta charset="UTF-8">
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <meta name="google-site-verification" content="S-FJQGrin5Z7C-YheOScSZRfMpd2wcrPb4pWS3L2zf0" />
+        <meta name="title" content="LockPick - NoPixel-MiniGames-4.0">
+        <meta name="description" content="Play the LockPick hacking minigame from NoPixel 4.0 for free">
+        <meta name="keywords" content="LockPick, NoPixel, MiniGames, 4.0, GTA, RP, GTARP, FiveM, Maximilian,">
         <link rel="stylesheet" type="text/css" href="../puzzle.css">
         <link rel="stylesheet" type="text/css" href="LockPick.css">
         <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:400,400i,700&display=swap">
         <script src="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/js/all.min.js"></script>
-        <script src="LockPick.js"></script>
+        <script type="module" src="LockPick.js"></script>
     </head>
     <body>
         <div class="other-container">
@@ -48,7 +51,14 @@
                     <i class="fas fa-hourglass-start"></i>
                     <p>Reset!</p>
                 </div>
-                <div class="hack-box"> 
+                <div class="hack-box">
+                    <div class="countdown-container">
+                        <div class="countdown">3</div>
+                        <div class="countdown two">3</div>
+                        <div class="countdown three">3</div>
+                        <div class="countdown four">3</div>
+                        <div class="countdown five">3</div>
+                    </div>
                     <div class="position-container">
                         <svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="100%" height="100%"></svg>
                     </div>
