+++ package.json
@@ -1,0 +1,13 @@
+{
+  "scripts": {
+    "build": "tsc -p .",
+    "serve": "serve ."
+  },
+  "devDependencies": {
+    "@types/node": "^20.11.19",
+    "serve": "^14.2.1"
+  },
+  "dependencies": {
+    "typescript": "^5.3.3"
+  }
+}
